server/requirements.txt
argon2-cffi==19.1.0
asn1crypto==0.24.0
astroid==2.1.0
awsebcli==3.14.6
bcrypt==3.1.6
botocore==1.12.15
cement==2.8.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.5
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.5.0
django-environ==0.4.5
django-heroku==0.3.1
django-sslserver==0.20
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.14
environ==1.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.2
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.16.0
pandas==0.24.0
paramiko==2.4.2
pathspec==0.5.5
pipreqs==0.4.9
psycopg2==2.7.6.1
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
PyNaCl==1.3.0
python-dateutil==2.7.3
pytz==2018.9
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
semantic-version==2.5.0
six==1.11.0
sklearn==0.0
termcolor==1.1.0
urllib3==1.22
whitenoise==4.1.2
wrapt==1.11.0
yarg==0.1.9
